requirements.txt
discord.py==2.4.0
aiohttp==3.9.5
feedparser==6.0.11
beautifulsoup4==4.12.3
python-dotenv==1.0.1
